+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Removed reference to Video Manager and Audio Manager from Sync Video Player
 - Video Manager and Audio Manager objects register themselves with a video player on startup
 - Updated sync video player integrity check to look for none or duplicate managers referencing it
+- Added menu items to GameObject menu to add different VideoTXL objects to scene
 
 ## [2.0.0] - 07-29-23